A Better Installer Improved Installer by allowing user to select install and search directory Started work on the client struct for holding data about the various clients

+++ makefile
@@ -14,5 +14,5 @@ $(EXE): $(OBJECTS)
 	$(CC) $(CFLAGS) $< -o $@
 
 clean:
-	rm $(EXE)
+	if [ -e $(EXE) ] ; then rm $(EXE); fi
 	rm -rf *.o
